lab5/examples/fullstack-app/backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
mysql-connector-python==8.1.0

